requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.1
telnyx==2.1.6
gunicorn==21.2.0
